pyproject.toml
[tool.poetry]
name = "data-sync"
version = "0.1.0"
description = ""
authors = ["Rich Kevan <rskevan@me.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
